refactor(tui): refactor event list key handler
